adds universal-ssr and server-ssr examples
