+++ svelte/sveltekit-ssr/package.json
@@ -1,40 +1,40 @@
 {
-    "name": "buf-svelte",
-    "version": "0.0.1",
-    "scripts": {
-        "start": "npm run buf:generate && vite dev --port 3000",
-        "build": "vite build",
-        "test": "npm run playwright:install && playwright test",
-        "playwright:install": "playwright install",
-        "buf:generate": "npx buf generate buf.build/bufbuild/eliza",
-        "preview": "vite preview",
-        "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
-        "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-        "lint": "prettier --check --ignore-path ../.prettierignore . && eslint .",
-        "format": "prettier --write --ignore-path ../.prettierignore ."
-    },
-    "devDependencies": {
-        "@bufbuild/buf": "^1.18.0-1",
-        "@bufbuild/connect-web": "^0.9.0",
-        "@bufbuild/protobuf": "^1.2.0",
-        "@bufbuild/protoc-gen-connect-es": "^0.9.0",
-        "@bufbuild/protoc-gen-es": "^1.2.0",
-        "@playwright/test": "^1.30.0",
-        "@sveltejs/adapter-auto": "next",
-        "@sveltejs/kit": "next",
-        "@typescript-eslint/eslint-plugin": "^5.59.2",
-        "@typescript-eslint/parser": "^5.54.0",
-        "eslint": "^8.39.0",
-        "eslint-config-prettier": "^8.8.0",
-        "eslint-plugin-svelte3": "^4.0.0",
-        "prettier": "^2.8.8",
-        "prettier-plugin-svelte": "^2.10.0",
-        "svelte": "^3.58.0",
-        "svelte-check": "^3.0.3",
-        "svelte-preprocess": "^5.0.1",
-        "tslib": "^2.5.0",
-        "typescript": "^4.8.4",
-        "vite": "^4.2.1"
-    },
-    "type": "module"
+  "name": "buf-svelte-sveltekit-ssr",
+  "version": "0.0.1",
+  "scripts": {
+    "start": "npm run buf:generate && vite dev --port 3000",
+    "build": "vite build",
+    "test": "npm run playwright:install && playwright test",
+    "playwright:install": "playwright install",
+    "buf:generate": "npx buf generate buf.build/bufbuild/eliza",
+    "preview": "vite preview",
+    "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+    "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
+    "lint": "prettier --check --ignore-path ../.prettierignore . && eslint .",
+    "format": "prettier --write --ignore-path ../.prettierignore ."
+  },
+  "devDependencies": {
+    "@bufbuild/buf": "^1.15.0-1",
+    "@bufbuild/connect-web": "^0.9.0",
+    "@bufbuild/protobuf": "^1.2.0",
+    "@bufbuild/protoc-gen-connect-es": "^0.9.0",
+    "@bufbuild/protoc-gen-es": "^1.2.0",
+    "@playwright/test": "^1.30.0",
+    "@sveltejs/adapter-auto": "next",
+    "@sveltejs/kit": "next",
+    "@typescript-eslint/eslint-plugin": "^5.59.2",
+    "@typescript-eslint/parser": "^5.54.0",
+    "eslint": "^8.39.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-svelte3": "^4.0.0",
+    "prettier": "^2.8.8",
+    "prettier-plugin-svelte": "^2.10.0",
+    "svelte": "^3.58.0",
+    "svelte-check": "^3.0.3",
+    "svelte-preprocess": "^5.0.1",
+    "tslib": "^2.5.0",
+    "typescript": "^4.8.4",
+    "vite": "^4.2.1"
+  },
+  "type": "module"
 }
